pyproject.toml
[tool.poetry]
name = "sprite_ai"
version = "1.0.0"
description = ""
authors = ["Caio D <caiobd@protonmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.13"
PyQt5 = "^5.15.7"
python-rofi = "^1.0.1"
pyperclip = "^1.8.2"
pydantic = "^2.4.2"
pyyaml = "^6.0.1"
langchain = "^0.0.327"
llama-cpp-python = "^0.2.12"
dbus-python = "^1.3.2"
plyer = "^2.1.0"
platformdirs = "^4.0.0"
progressbar2 = "^4.2.0"
loguru = "^0.7.2"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
pytest = "^7.2.0"
mypy = "^0.991"
pyinstaller = "^6.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sprite-ai = "sprite_ai.main:main"

[repositories]
[repositories.my-gitlab]
url = "https://gitlab.com/api/v4/projects/<your project id>/packages/pypi"